+++ src/values.rs
@@ -26,7 +26,7 @@ pub enum Value {
 
     /// A reference to an object allocated in the 
     /// package elements vector
-    PackageRef(usize), 
+    PackageRef(i16),
